Menu with breadcrumbs v2 + refactor package
