--- src/main/java/com/infoshareacademy/MenuBreadcrumbs.java
@@ -1,5 +1,0 @@
-package com.infoshareacademy;
-
-public class MenuBreadcrumbs {
-    public String breadcrumbs= "";
-}
