+++ src/main/java/com/isa/zajavieni/menu/MainMenu.java
@@ -1,6 +1,5 @@
 package com.isa.zajavieni.menu;
 
-import com.infoshareacademy.BreadcrumbHistory;
 import com.isa.zajavieni.service.PrinterEvents;
 import com.isa.zajavieni.repository.EventList;
 import com.isa.zajavieni.service.SearchEvent;
